// src/types/topicSearch.ts
// Generated by https://quicktype.io

export interface TopicSearchResponse {
  total_count: number;
  incomplete_results: boolean;
  items: Item[];
}

export interface Item {
  name: string;
  display_name: null | string;
  short_description: null | string;
  description: null | string;
  created_by: null | string;
  released: null | string;
  created_at: string;
  updated_at: string;
  featured: boolean;
  curated: boolean;
  score: number;
}